kvstore: Move dummy logic to test file
